Replace System.IO.FileSystem Directory_SetLastAccessTime_str_dt tests
