+++ src/System.IO.FileSystem/tests/System.IO.FileSystem.Tests.csproj
@@ -37,6 +37,8 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Rewritten -->
+    <Compile Include="Directory\SetLastWriteTime_str_dt.cs" />
+    <Compile Include="Directory\SetLastAccessTime_str_dt.cs" />
     <Compile Include="DirectoryInfo\Create.cs" />
     <Compile Include="DirectoryInfo\CreateSubdirectory_str.cs" />
     <Compile Include="DirectoryInfo\ctor_str.cs" />
@@ -81,6 +83,7 @@
     <Compile Include="UnseekableFileStream.cs" />
     <Compile Include="FSAssert.cs" />
     <Compile Include="FileSystemTest.cs" />
+    <Compile Include="SetTime_str_dt.cs" />
     <!-- Ported -->
     <Compile Include="PortedCommon\CommonUtilities.cs" />
     <Compile Include="PortedCommon\DllImports.cs" />
@@ -130,8 +133,6 @@
     <Compile Include="Directory\ReparsePoints_MountVolume.cs" />
     <Compile Include="Directory\SetCreationTime_str_dt.cs" />
     <Compile Include="Directory\SetCurrentDirectory_dir.cs" />
-    <Compile Include="Directory\SetLastAccessTime_str_dt.cs" />
-    <Compile Include="Directory\SetLastWriteTime_str_dt.cs" />
     <Compile Include="Directory\Ansichars.cs" />
     <Compile Include="Directory\CreateDirectory.cs" />
     <Compile Include="Directory\Delete_MountVolume.cs" />
